+++ Assets/lilToon/Editor/lilInspector.cs
@@ -3611,6 +3611,7 @@ namespace lilToon
             #endif
 
             EditorGUI.BeginChangeCheck();
+            ToggleGUI("Optimize in NDMF (Apply on Play)", ref shaderSetting.isOptimizeInNDMF);
             ToggleGUI(GetLoc("sShaderSettingOptimizeInEditor"), ref shaderSetting.isDebugOptimize);
             ToggleGUI("Migrate materials in startup", ref shaderSetting.isMigrateInStartUp);
             edSet.isShowShaderSetting = lilEditorGUI.Foldout(GetLoc("sShaderSetting"), edSet.isShowShaderSetting);