+++ Assets/lilToon/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Shader optimization can now be skipped in NDMF's `Apply on Play`
+
 ### Fixed
 - Fur rim light calculations may be incorrect after optimization
 - UV Tile Discard not working properly